Move common mock methods to helper class
